Task3

+++ src/main/java/pw/react/backend/reactbackend/service/UserServiceImpl.java
@@ -10,21 +10,26 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import pw.react.backend.reactbackend.dao.UserDao;
 import pw.react.backend.reactbackend.entity.User;
+import pw.react.backend.reactbackend.repository.FindUserRepo;
 
 @Service
 public class UserServiceImpl implements UserService{
-	@Resource 
+	@Resource
 	UserDao userDao;
 	@Resource
 	UserExistsService userExistsService;
+	@Resource
+	FindUserRepo findUserRepo;
 	@Override
-	public List<User> findAll() {
-		return userDao.findAll();
+	public Object findUser(String login) {
+		User usr = findUserRepo.find(login);
+		if (usr==null) return new ResponseEntity("User with such login is not present in the database.", HttpStatus.CONFLICT);
+		else return usr;
 	}
 	@Override
 	public ResponseEntity insertUser(User usr) {
 		if (userExistsService.exists(usr))
-			return new ResponseEntity("User with such a login is already present in database.", HttpStatus.CONFLICT);
+			return new ResponseEntity("User with such login is already present in the database.", HttpStatus.CONFLICT);
 		else
 		{
 			userDao.insertUser(usr);
@@ -32,19 +37,25 @@ public class UserServiceImpl implements UserService{
 		}
 	}
 	@Override
-	public void updateUser(User usr) {
-		userDao.updateUser(usr);
-		
+	public ResponseEntity updateUser(User usr) {
+		if (userExistsService.exists(usr)) {
+			userDao.updateUser(usr);
+			return new ResponseEntity(HttpStatus.OK);
+		}
+		else return new ResponseEntity("User with such a login is not present in the database.", HttpStatus.CONFLICT);
 	}
 	@Override
 	public void executeUpdateUser(User usr) {
 		userDao.executeUpdateUser(usr);
-		
 	}
 
 	@Override
-	public void deleteUser(User usr) {
-		userDao.deleteUser(usr);
-		
+	public ResponseEntity deleteUser(String login) {
+		User usr = findUserRepo.find(login);
+		if (usr==null) return new ResponseEntity("User with such a login is not present in the database.", HttpStatus.CONFLICT);
+		else {
+			userDao.deleteUser(usr);
+			return new ResponseEntity(HttpStatus.OK);
+		}
 	}
 }
